backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-openai
pinecone
pypdf
tiktoken
python-dotenv
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
pytest
pytest-asyncio
httpx
email-validator
